pyproject.toml
[tool.poetry]
name = "metamaker"
version = "0.1.0"
description = ""
authors = ["altescy <altescy@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.19.9"
sagemaker = {version = "^2.68.0", extras = ["local"]}
uvicorn = "^0.15.0"
fastapi = "^0.70.0"
pyyaml = ">=5.0,<6.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
python-language-server = "^0.36.2"
pysen = {version = "^0.9.1", extras = ["lint"]}
pytest = "^6.2.4"
boto3-stubs = {version = "^1.19.9", extras = ["ecr", "s3", "sagemaker"]}

[tool.poetry.scripts]
metamaker = "metamaker.__main__:run"

[tool.pysen]
version = "0.9"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 119
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pysen.lint.source]
  includes = ["."]
  excludes = [".venv/"]

[tool.black] # automatically generated by pysen
line-length = 119
target-version = ["py38"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 119
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
